backend/pyproject.toml
[tool.poetry]
name = "onlybeer"
version = "0.0.2"
description = "OnlyBeer - Api"
authors = ["Gerardo Gonzalez <gerardmga@gmail.com>"]
packages = [
    { include = "onlybeer" },
]

[tool.poetry.dependencies]
# Dependencies for core services
python = "3.10.11"
django = "5.0.6"
django-filter = "22.1"
django_extensions = "3.2.1"
djangorestframework = "3.15.0"
djangorestframework-simplejwt = "5.2.2"
drf-yasg = "1.21.5"
asgiref = "^3.7.0"
gunicorn = "^20.1.0"
pika = "1.3.2"
pillow = "9.4.0"
coverage= "7.2.1"
django-cors-headers = "^4.3.1"

[tool.poetry.dev-dependencies]
django-debug-toolbar= "^3.8.1"
